requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
firebase-admin==6.2.0
python-dotenv==1.0.0
Pillow==10.0.1
pytest==7.4.2
gunicorn==21.2.0
flask-jwt-extended==4.5.3
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
marshmallow==3.20.1
tensorflow==2.13.0
numpy==1.24.3
pandas==2.1.1